Реализация ФЗ12 2024.10.31

+++ src/main/java/ru/yandex/practicum/filmorate/dal/mappers/FilmRowMapper.java
@@ -13,7 +13,7 @@ public class FilmRowMapper implements RowMapper<Film> {
     @Override
     public Film mapRow(ResultSet resultSet, int rowNum) throws SQLException {
         Film film = new Film();
-        film.setId(resultSet.getInt("id"));
+        film.setId(resultSet.getLong("id"));
         film.setName(resultSet.getString("name"));
         film.setDescription(resultSet.getString("description"));
         film.setReleaseDate(resultSet.getTimestamp("release_date").toLocalDateTime().toLocalDate());
